added mapping class after manual edit
