+++ scripts/oscal_normalize.py
@@ -83,13 +83,14 @@ class_header = 'class '
 
 # List of filestems not including 'complete' or 'common'
 # 'common' is generated by this script.  'complete.py' comes from NIST and is ignored
-fstems = ['assessment_plan', 'assessment_results', 'catalog', 'component', 'poam', 'profile', 'ssp']
+fstems = ['assessment_plan', 'assessment_results', 'catalog', 'component', 'mapping', 'poam', 'profile', 'ssp']
 
 alias_map = {
     'assessment_plan': 'assessment-plan',
     'assessment_results': 'assessment-results',
     'catalog': 'catalog',
     'component': 'component-definition',
+    'mapping': 'mapping',
     'poam': 'plan-of-action-and-milestones',
     'profile': 'profile',
     'ssp': 'system-security-plan'
@@ -100,6 +101,7 @@ camel_map = {
     'assessment_results': 'AssessmentResults',
     'catalog': 'Catalog',
     'component': 'ComponentDefinition',
+    'mapping': 'Mapping',
     'poam': 'PlanOfActionAndMilestones',
     'profile': 'Profile',
     'ssp': 'SystemSecurityPlan'
@@ -110,6 +112,7 @@ prefix_map = {
     'assessment_results': 'Ar',
     'catalog': 'Cat',
     'component': 'Comp',
+    'mapping': 'Map',
     'poam': 'Poam',
     'profile': 'Prof',
     'ssp': 'Ssp'
@@ -164,7 +167,7 @@ from __future__ import annotations
 import re
 from datetime import datetime
 from enum import Enum
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, List, Optional, Union
 
 from pydantic import AnyUrl, EmailStr, Extra, Field, conint, constr, validator
